refactor tableview api
