+++ development/tests/tableview.py
@@ -2,6 +2,7 @@ import ttkbootstrap as ttk
 from ttkbootstrap.constants import *
 from pathlib import Path
 import csv
+from ttkbootstrap.tableview import Tableview
 from ttkbootstrap.utility import scale_size
 
 app = ttk.Window(themename='sandstone')
@@ -24,37 +25,7 @@ coldata = [
     {"text": "Leave", "stretch": False},
 ]
 
-# tv = ttk.Treeview(bootstyle=PRIMARY, columns = [0, 1, 2, 3, 4], show=HEADINGS)
-# colnames = ['SerialNumber', 'CompanyName', 'Employee', 'Description', 'Leave']
-# for i, col in enumerate(colnames):
-#     tv.heading(i, text=col, anchor=W)
-#     tv.column(i, stretch=False)
-
-# for row in rowdata:
-#     tv.insert('', END, values=row)
-
-# tv.pack(fill=BOTH, expand=YES, padx=10, pady=(10, 0))
-# sb = ttk.Scrollbar(orient=HORIZONTAL, command=tv.xview)
-# tv.configure(xscrollcommand=sb.set)
-# sb.pack(fill=X, padx=10, pady=(0, 10))
-
-# # calculate the best size for each column
-# from tkinter import font
-# f = font.nametofont('TkDefaultFont')
-# widths = []
-# for col in colnames:
-#     widths.append(f.measure(col) + scale_size(tv, 10))
-
-# for row in rowdata:
-#     for i, col in enumerate(row):
-#         measure = f.measure(col)
-#         if measure > widths[i]:
-#             widths[i] = measure
-
-# for i, width in enumerate(widths):
-#     tv.column(i, width=width)
-
-dt = ttk.Tableview(
+dt = Tableview(
     master=app,
     coldata=coldata,
     rowdata=rowdata,
@@ -65,6 +36,5 @@ dt = ttk.Tableview(
     autofit=False
 )
 dt.pack(fill=BOTH, expand=YES, padx=5, pady=5)
-print(app.style.colors.inputbg)
 
 app.mainloop()
