+++ docs/api/tableview.md
@@ -1,8 +1,7 @@
-# Tableview
+# tableview module
 
-::: ttkbootstrap.widgets.Tableview
+::: ttkbootstrap.tableview
     selection:
         filters: ["!^_", "^__init__"]
     rendering:
         heading_level: 2
-        show_root_heading: true
