+++ src/ttkbootstrap/widgets.py
@@ -1039,1609 +1039,3 @@ class Meter(ttk.Frame):
             self.amountusedvar.set(amountused - delta)
         else:
             self.amountusedvar.set(amountused + delta)
-
-
-UPARROW = "⬆"
-DOWNARROW = "⬇"
-ASCENDING = 0
-DESCENDING = 1
-
-
-class Tableview(ttk.Frame):
-    """A class for arranging data in rows and columns. A Tableview
-    object contains various features such has striped rows, pagination,
-    and autosized and autoaligned columns.
-
-    The pagination option is recommended when loading a lot of data as
-    the table records are inserted on-demand. Table records are only
-    created when requested to be in a page view. This allows the table
-    to be loaded very quickly even with hundreds of thousands of
-    records.
-
-    All table columns are sortable. Clicking a column header will toggle
-    between sorting "ascending" and "descending".
-
-    Columns are configurable by passing a simple list of header names or
-    by passing in a dictionary of column names with settings. You can
-    use both as well, as in the example below, where a column header
-    name is use for one column, and a dictionary of settings is used
-    for another.
-
-    The object has a right-click menu on the header and the cells that
-    allow you to configure various settings.
-
-    ![](../../assets/widgets/tableview-1.png)
-    ![](../../assets/widgets/tableview-2.png)
-
-    Examples:
-
-        ```python
-        import ttkbootstrap as ttk
-        from ttkbootstrap.constants import *
-
-        app = ttk.Window()
-        colors = app.style.colors
-
-        coldata = [
-            {"text": "LicenseNumber", "stretch": False},
-            "CompanyName",
-            {"text": "UserCount", "stretch": False},
-        ]
-
-        rowdata = [
-            ('A123', 'IzzyCo', 12),
-            ('A136', 'Kimdee Inc.', 45),
-            ('A158', 'Farmadding Co.', 36)
-        ]
-
-        dt = ttk.Tableview(
-            master=app,
-            coldata=coldata,
-            rowdata=rowdata,
-            paginated=True,
-            searchable=True,
-            bootstyle=PRIMARY,
-            stripecolor=(colors.light, None),
-        )
-        dt.pack(fill=BOTH, expand=YES, padx=10, pady=10)
-
-        app.mainloop()
-        ```
-    """
-
-    def __init__(
-        self,
-        master=None,
-        bootstyle=DEFAULT,
-        coldata=[],
-        rowdata=[],
-        paginated=False,
-        searchable=False,
-        autofit=False,
-        autoalign=True,
-        stripecolor=None,
-        pagesize=10,
-        height=10,
-    ):
-        """
-        Parameters:
-
-            master (Widget):
-                The parent widget.
-
-            bootstyle (str):
-                A style keyword used to set the focus color of the entry
-                and the background color of the date button. Available
-                options include -> primary, secondary, success, info,
-                warning, danger, dark, light.
-
-            coldata (List[str | Dict]):
-                An iterable containing either the heading name or a
-                dictionary of column settings. Configurable settings
-                include >> text, image, command, anchor, width, minwidth,
-                maxwidth, stretch
-
-            rowdata (List):
-                An iterable of row data. The lenth of each row of data
-                must match the number of columns.
-
-            paginated (bool):
-                Specifies that the data is to be paginated. A pagination
-                frame will be created below the table with controls that
-                enable the user to page forward and backwards in the
-                data set.
-
-            pagesize (int):
-                When `paginated=True`, this specifies the number of rows
-                to show per page. 
-
-            searchable (bool):
-                If `True`, a searchbar will be created above the table.
-                Press the <Return> key to initiate a search. Searching
-                with an empty string will reset the search criteria, or
-                pressing the reset button to the right of the search
-                bar. Currently, the search method looks for any row
-                that contains the search text. The filtered results
-                are displayed in the table view.
-
-            autofit (bool):
-                If `True`, the table columns will be automatically sized
-                when loaded based on the records in the current view.
-
-            autoalign (bool):
-                If `True`, the column headers and data are automatically
-                aligned. Numbers and number headers are right-aligned
-                and all other data types are left-aligned. The auto
-                align method evaluates the first record in each column
-                to determine the data type for alignment.
-
-            stripecolor (Tuple[str, str]):
-                If provided, even numbered rows will be color using the
-                (background, foreground) specified. You may specify one
-                or the other by passing in **None**. For example,
-                `stripecolor=('green', None)` will set the stripe
-                background as green, but the foreground will remain as
-                default. You may use standand color names, hexadecimal
-                color codes, or bootstyle color keywords. For example,
-                ('light', '#222') will set the background to the "light"
-                themed ttkbootstrap color and the foreground to the
-                specified hexadecimal color.
-
-            height (int):
-                Specifies how many rows will appear in the table's viewport.
-                If the number of records extends beyond the table height,
-                the user may use the mousewheel or scrollbar to navigate
-                the data.
-        """
-        super().__init__(master)
-        self.tablecols = []
-        self.tablerows = []
-        self.tablerows_filtered = []
-        self.viewdata = []
-        self.rowindex = tk.IntVar(value=0)
-        self.pageindex = tk.IntVar(value=1)
-        self.pagelimit = tk.IntVar(value=0)
-        self.height = height
-        self.pagesize = tk.IntVar(value=pagesize)
-        self.paginated = paginated
-        self.searchable = searchable
-        self.stripecolor = stripecolor
-        self.autofit = autofit
-        self.autoalign = autoalign
-        self.filtered = False
-        self.criteria = tk.StringVar()
-        self.rightclickmenu_cell = None
-
-        self._build_table(coldata, rowdata, bootstyle)
-
-    # DATA LOADING
-
-    def unload_table_data(self):
-        """Unload all data from the table"""
-        for row in self.viewdata:
-            row.hide()
-        self.viewdata.clear()
-
-    def load_table_data(self):
-        """Load all data into the table"""
-        self.unload_table_data()
-        page_start = self.rowindex.get()
-        page_end = self.rowindex.get() + self.pagesize.get()
-
-        if self.filtered:
-            rowdata = self.tablerows_filtered[page_start:page_end]
-            rowcount = len(self.tablerows_filtered)
-        else:
-            rowdata = self.tablerows[page_start:page_end]
-            rowcount = len(self.tablerows)
-
-        self.pagelimit.set(ceil(rowcount / self.pagesize.get()))
-
-        pageindex = ceil(page_end / self.pagesize.get())
-        pagelimit = self.pagelimit.get()
-        self.pageindex.set(min([pagelimit, pageindex]))
-
-        for i, row in enumerate(rowdata):
-            if self.stripecolor is not None and i % 2 == 0:
-                row.show(True)
-            else:
-                row.show(False)
-            self.viewdata.append(row)
-
-    def _get_event_objects(self, event):
-        iid = self.tableview.identify_row(event.y)
-        item = self.tableview.item(iid)
-        col = self.tableview.identify_column(event.x)
-        cid = int(self.tableview.column(col, "id"))
-        values = item.get("values")
-        data = {
-            'iid': iid,
-            'cid': cid,
-            'col': col,
-            'item': item,
-            'values': values,
-        }
-        if values:
-            data['value'] = values[cid]
-        return data
-
-    # WIDGET BUILDERS
-
-    def _build_table(self, coldata, rowdata, bootstyle):
-        """Build the data table"""
-        if self.searchable:
-            self._build_search_frame()
-
-        self.tableview = ttk.Treeview(
-            master=self,
-            columns=[x for x in range(len(coldata))],
-            height=self.height,
-            selectmode=EXTENDED,
-            show=HEADINGS,
-            bootstyle=f"{bootstyle}-table",
-        )
-        self.tableview.pack(fill=BOTH, expand=YES, side=TOP)
-        self.hbar = ttk.Scrollbar(
-            master=self, 
-            command=self.tableview.xview, 
-            orient=HORIZONTAL
-        )
-        self.hbar.pack(fill=X)
-        self.tableview.configure(xscrollcommand=self.hbar.set)
-
-        self._build_table_columns(coldata)
-        self._build_table_rows(rowdata)
-        # self.build_horizontal_scrollbar()
-
-        self.load_table_data()
-
-        if self.autofit:
-            self.autofit_columns()
-
-        if self.autoalign:
-            self.autoalign_columns()
-
-        if self.paginated:
-            self._build_pagination_frame()
-
-        if self.stripecolor is not None:
-            self.configure_table_stripes(self.stripecolor)
-
-        self.rightclickmenu_cell = TableCellRightClickMenu(self)
-        self.rightclickmenu_head = TableHeaderRightClickMenu(self)
-        self._set_widget_binding()
-
-    def _build_search_frame(self):
-        """Build the search frame containing the search widgets. This
-        frame is only created if `searchable=True` when creating the
-        widget.
-        """
-        frame = ttk.Frame(self, padding=5)
-        frame.pack(fill=X, side=TOP)
-        ttk.Label(frame, text="Search").pack(side=LEFT, padx=5)
-        searchterm = ttk.Entry(frame, textvariable=self.criteria)
-        searchterm.pack(fill=X, side=LEFT, expand=YES)
-        searchterm.bind("<Return>", self._search_table_data)
-        searchterm.bind("<KP_Enter>", self._search_table_data)
-        ttk.Button(
-            frame,
-            text="⎌",
-            command=self.clear_table_filters,
-            style="symbol.Link.TButton",
-        ).pack(side=LEFT)
-
-    def _build_pagination_frame(self):
-        """Build the frame containing the pagination widgets. This
-        frame is only built if `pagination=True` when creating the
-        widget.
-        """
-        pageframe = ttk.Frame(self)
-        pageframe.pack(fill=X, anchor=N)
-
-        ttk.Button(
-            master=pageframe,
-            text="»",
-            command=self.goto_last_page,
-            style="symbol.Link.TButton",
-        ).pack(side=RIGHT, fill=Y)
-        ttk.Button(
-            master=pageframe,
-            text="›",
-            command=self.goto_next_page,
-            style="symbol.Link.TButton",
-        ).pack(side=RIGHT, fill=Y)
-
-        ttk.Separator(pageframe, orient=VERTICAL).pack(side=RIGHT)
-        lbl = ttk.Label(pageframe, textvariable=self.pagelimit)
-        lbl.pack(side=RIGHT, padx=(0, 5))
-        ttk.Label(pageframe, text="of").pack(side=RIGHT, padx=(5, 0))
-
-        index = ttk.Entry(pageframe, textvariable=self.pageindex, width=4)
-        index.pack(side=RIGHT)
-        index.bind("<Return>", self.goto_page, "+")
-        index.bind("<KP_Enter>", self.goto_page, "+")
-
-        ttk.Label(pageframe, text="Page").pack(side=RIGHT, padx=5)
-        ttk.Separator(pageframe, orient=VERTICAL).pack(side=RIGHT)
-
-        ttk.Button(
-            master=pageframe,
-            text="‹",
-            command=self.goto_prev_page,
-            style="symbol.Link.TButton",
-        ).pack(side=RIGHT, fill=Y)
-        ttk.Button(
-            master=pageframe,
-            text="«",
-            command=self.goto_first_page,
-            style="symbol.Link.TButton",
-        ).pack(side=RIGHT, fill=Y)
-
-        ttk.Label(pageframe, text="Pagesize").pack(side=LEFT, padx=5, fill=Y)
-        values = [5, 10, 25, 50, 75, 100]
-        cbo = ttk.Combobox(
-            master=pageframe, 
-            values=values, 
-            textvariable=self.pagesize,
-            width=4,
-            state=READONLY
-        )
-        cbo.pack(side=LEFT)
-        cbo.bind("<<ComboboxSelected>>", self._select_pagesize)
-
-    def _build_table_rows(self, rowdata):
-        """Build, load, and configure the DataTableRow objects
-
-        Parameters:
-
-            rowdata (List):
-                An iterable of row data
-        """
-        for row in rowdata:
-            self.tablerows.append(TableRow(self.tableview, row))
-
-    def _build_table_columns(self, coldata):
-        """Build, load, and configure the DataTableColumn objects
-
-        Parameters:
-
-            coldata (List[str|Dict[str, Any]]):
-                An iterable of column names or a dictionary of column
-                configuration settings.
-        """
-        for cid, col in enumerate(coldata):
-            if isinstance(col, str):
-                self.tablecols.append(
-                    TableColumn(
-                        table=self.tableview,
-                        cid=cid,
-                        text=col,
-                    )
-                )
-            else:
-                if "text" not in col:
-                    col["text"] = f"Column {cid}"
-                self.tablecols.append(
-                    TableColumn(
-                        table=self.tableview,
-                        cid=cid,
-                        **col,
-                    )
-                )
-
-    # PAGE NAVIGATION
-
-    def goto_first_page(self):
-        """Update table with first page of data"""
-        self.rowindex.set(0)
-        self.load_table_data()
-
-    def goto_last_page(self):
-        """Update table with the last page of data"""
-        self.rowindex.set(self.pagesize.get() * (self.pagelimit.get() - 1))
-        self.load_table_data()
-
-    def goto_next_page(self):
-        """Update table with next page of data"""
-        if self.pageindex.get() >= self.pagelimit.get():
-            return
-        rowindex = self.rowindex.get()
-        self.rowindex.set(rowindex + self.pagesize.get())
-        self.load_table_data()
-
-    def goto_prev_page(self):
-        """Update table with prev page of data"""
-        if self.pageindex.get() <= 1:
-            return
-        rowindex = self.rowindex.get()
-        self.rowindex.set(rowindex - self.pagesize.get())
-        self.load_table_data()
-
-    def goto_page(self, *_):
-        """Go to a specific page indicated by the page entry widget."""
-        pageindex = self.pageindex.get() - 1
-        self.rowindex.set(pageindex * self.pagesize.get())
-        self.load_table_data()
-
-    # COLUMN SORTING
-
-    def sort_column_data(self, event=None, cid=None, sort=None):
-        """Sort the table rows by the specified column. This method 
-        may be trigged by an event or manually.
-        
-        Parameters:
-
-            event (Event):
-                A window event.
-
-            cid (int):
-                A unique column identifier; typically the numerical
-                index of the column relative to the original data set.
-
-            sort (int):
-                Determines the sort direction. 0 = ASCENDING. 1 = DESCENDING.
-        """
-        if event is not None:
-            eo = self._get_event_objects(event)
-            cid = eo.get('cid')
-        elif cid is None:
-            return
-
-        # update table data
-        if self.filtered:
-            tablerows = self.tablerows_filtered
-        else:
-            tablerows = self.tablerows
-
-        if sort is not None:
-            colsort = sort
-        else:
-            colsort = self.tablecols[cid].sort
-
-        if colsort == ASCENDING:
-            self.tablecols[cid].sort = DESCENDING
-        else:
-            self.tablecols[cid].sort = ASCENDING
-
-        sortedrows = sorted(
-            tablerows, reverse=colsort, key=lambda x: x.values[cid]
-        )
-        if self.filtered:
-            self.tablerows_filtered = sortedrows
-        else:
-            self.tablerows = sortedrows
-
-        # update headers
-        self._column_sort_header_reset()
-        self._column_sort_header_update(cid)
-
-        self.unload_table_data()
-        self.load_table_data()
-
-    def _column_sort_header_reset(self):
-        """Remove the sort character from the column headers"""
-        for col in self.tablecols:
-            self.tableview.heading(col.cid, text=col.headertext)
-
-    def _column_sort_header_update(self, cid):
-        """Add sort character to the sorted column"""
-        col = self.tablecols[cid]
-        arrow = UPARROW if col.sort == ASCENDING else DOWNARROW
-        headertext = f"{col.headertext} {arrow}"
-        self.tableview.heading(col.cid, text=headertext)
-
-    # DATA SEARCH & FILTERING
-
-    def clear_table_filters(self):
-        """Remove all filters from table data and reset the filtered 
-        flag.
-        """
-        self.filtered = False
-        self.criteria.set("")
-        self.unload_table_data()
-        self.load_table_data()
-        self._column_sort_header_reset()
-
-    def filter_column_to_value(self, event=None, cid=None, value=None):
-        """Hide all records except for records where the current
-        column exactly matches the provided value. This method may 
-        be triggered by a window event or by specifying the column id.
-        
-        Parameters:
-
-            event (Event):
-                A window event.
-
-            cid (int):
-                A unique column identifier; typically the numerical
-                index of the column within the original dataset.
-
-            value (Any):
-                The criteria used to filter the column.
-        """
-        if event is not None:
-            eo = self._get_event_objects(event)
-            cid = eo.get('cid')
-            value = eo.get('value') or value
-        elif cid is None:
-            return
-        self.filtered = True
-        self.tablerows_filtered.clear()
-        self.unload_table_data()
-        data = self.tablerows
-        for row in data:
-            if row.values[cid] == value:
-                self.tablerows_filtered.append(row)
-        self.rowindex.set(0)
-        self.load_table_data()
-
-    def filter_to_selected_rows(self):
-        """Hide all records except for the selected rows"""
-        criteria = self.tableview.selection()
-        if len(criteria) == 0:
-            return # nothing is selected
-
-        if self.filtered:
-            for row in self.viewdata:
-                if row.iid not in criteria:
-                    row.hide()
-                    self.tablerows_filtered.remove(row)
-        else:
-            self.filtered = True
-            self.tablerows_filtered.clear()
-            for row in self.viewdata:
-                if row.iid in criteria:
-                    self.tablerows_filtered.append(row)
-        self.rowindex.set(0)
-        self.load_table_data()
-
-    def hide_selected_rows(self):
-        """Hide the currently selected rows"""
-        selected = self.tableview.selection()
-        self.tableview.detach(*selected)
-        tablerows = [row for row in self.viewdata if row.iid in selected]
-
-        if not self.filtered:
-            self.filtered = True
-            self.tablerows_filtered = self.tablerows
-
-        for row in tablerows:
-            if self.filtered:
-                self.tablerows_filtered.remove(row)
-
-        self.load_table_data()
-
-    def hide_selected_column(self, event=None, cid=None):
-        """Detach the selected column from the tableview. This method
-        may be triggered by a window event or by specifying the column
-        id.
-        
-        Parameters:
-
-            event (Event):
-                A window event.
-
-            cid (int):
-                A unique column identifier; typically the numerical
-                index of the column within the original dataset.
-        """
-        if event is not None:
-            eo = self._get_event_objects(event)
-            cid = eo.get('cid')
-        elif cid is None:
-            return
-        displaycols = list(self.tableview.configure("displaycolumns"))
-        cols = list(self.tableview.configure("columns"))
-        if "#all" in displaycols:
-            displaycols = cols
-        displaycols.remove(str(cid))
-        self.tableview.configure(displaycolumns=displaycols)
-
-    def show_selected_column(self, event=None, cid=None):
-        """Attach the selected column to the tableview. This method
-        may be triggered by a window event or by specifying the column
-        id. The column is reinserted at the index in the original data
-        set.
-        
-        Parameters:
-
-            event (Event):
-                A window event.
-
-            cid (int):
-                A unique column identifier; typically the numerical
-                index of the column within the original dataset.
-        """
-        if event is not None:
-            eo = self._get_event_objects(event)
-            cid = eo.get('cid')
-        elif cid is None:
-            return
-        displaycols = list(self.tableview.configure("displaycolumns"))
-        if "#all" in displaycols:
-            return
-        if str(cid) in displaycols:
-            return
-        columns = list(self.tableview.configure("columns"))        
-        index = columns.index(str(cid))
-        displaycols.insert(index, str(cid))
-        self.tableview.configure(displaycolumns=displaycols)
-
-    def _search_table_data(self, _):
-        """Search the table data for records that meet search criteria.
-        Currently, this search locates any records that contain the
-        specified text; it is also case insensitive.
-        """
-        criteria = self.criteria.get()
-        self.filtered = True
-        self.tablerows_filtered.clear()
-        self.unload_table_data()
-        data = self.tablerows
-        for row in data:
-            for col in row.values:
-                if str(criteria).lower() in str(col).lower():
-                    self.tablerows_filtered.append(row)
-                    break
-        self.rowindex.set(0)
-        self.load_table_data()
-
-    # DATA EXPORT
-
-    def export_all_records(self):
-        """Export all records to a csv file"""
-        headers = [col.headertext for col in self.tablecols]
-        records = [row.values for row in self.tablerows]
-        self.save_data_to_csv(headers, records)
-
-    def export_current_page(self):
-        """Export records on current page to csv file"""
-        headers = [col.headertext for col in self.tablecols]
-        records = [row.values for row in self.viewdata]
-        self.save_data_to_csv(headers, records)
-
-    def export_current_selection(self):
-        """Export rows currently selected to csv file"""
-        headers = [col.headertext for col in self.tablecols]
-        selected = self.tableview.selection()
-        records = []
-        for iid in selected:
-            records.append(self.tableview.item(iid)["values"])
-        self.save_data_to_csv(headers, records)
-
-    def export_records_in_filter(self):
-        """Export rows currently filtered to csv file"""
-        headers = [col.headertext for col in self.tablecols]
-        if not self.filtered:
-            return
-        records = [row.values for row in self.tablerows_filtered]
-        self.save_data_to_csv(headers, records)
-
-    def save_data_to_csv(self, headers, records):
-        """Save data records to a csv file.
-
-        Parameters:
-
-            headers (List[str]):
-                A list of header labels.
-
-            records (List[Tuple[...]]):
-                A list of table records.
-        """
-        from tkinter.filedialog import asksaveasfilename
-        import csv
-
-        timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-        initialfile = f"tabledata_{timestamp}.csv"
-        filetypes = [
-            ("CSV UTF-8 (Comma delimited)", "*.csv"),
-            ("All file types", "*.*"),
-        ]
-        filename = asksaveasfilename(
-            confirmoverwrite=True,
-            filetypes=filetypes,
-            defaultextension="csv",
-            initialfile=initialfile,
-        )
-        if filename:
-            with open(filename, "w", encoding="utf-8", newline="") as f:
-                writer = csv.writer(f)
-                writer.writerow(headers)
-                writer.writerows(records)
-
-    # ROW MOVEMENT
-
-    def move_selected_rows_to_top(self):
-        """Move the selected rows to the top of the data set"""
-        selected = self.tableview.selection()
-        if len(selected) == 0:
-            return
-
-        if self.filtered:
-            tablerows = self.tablerows_filtered
-        else:
-            tablerows = self.tablerows
-
-        row_items = []       
-        
-        for iid in selected:
-            for row in tablerows:
-                if row.iid == iid:
-                    row_items.append(row)
-                    break
-
-        for i, item in enumerate(row_items):
-            tablerows.remove(item)
-            tablerows.insert(i, item)
-
-        if self.filtered:
-            self.tablerows_filtered = tablerows
-        else:
-            self.tablerows = tablerows
-
-        # refresh the table data
-        self.unload_table_data()
-        self.load_table_data()
-
-    def move_selected_rows_to_bottom(self):
-        """Move the selected rows to the bottom of the dataset"""
-        selected = self.tableview.selection()
-        if len(selected) == 0:
-            return
-
-        if self.filtered:
-            tablerows = self.tablerows_filtered
-        else:
-            tablerows = self.tablerows
-
-        row_items = []       
-        
-        for iid in selected:
-            for row in tablerows:
-                if row.iid == iid:
-                    row_items.append(row)
-                    break
-
-        for item in row_items:
-            tablerows.remove(item)
-            tablerows.append(item)
-
-        if self.filtered:
-            self.tablerows_filtered = tablerows
-        else:
-            self.tablerows = tablerows
-
-        # refresh the table data
-        self.unload_table_data()
-        self.load_table_data()
-
-    def move_selected_row_up(self):
-        """Move the selected rows up one position in the dataset"""
-        selected = self.tableview.selection()
-        if len(selected) == 0:
-            return
-
-        if self.filtered:
-            tablerows = self.tablerows_filtered
-        else:
-            tablerows = self.tablerows
-
-        row_items = []       
-        
-        for iid in selected:
-            for i, row in enumerate(tablerows):
-                if row.iid == iid:
-                    row_items.append([i, row])
-                    break
-
-        for index, item in row_items:
-            tablerows.remove(item)
-            tablerows.insert(index-1, item)
-
-        if self.filtered:
-            self.tablerows_filtered = tablerows
-        else:
-            self.tablerows = tablerows
-
-        # refresh the table data
-        self.unload_table_data()
-        self.load_table_data()
-
-    def move_row_down(self):
-        """Move the selected rows down one position in the dataset"""
-        selected = self.tableview.selection()
-        if len(selected) == 0:
-            return
-
-        if self.filtered:
-            tablerows = self.tablerows_filtered
-        else:
-            tablerows = self.tablerows
-
-        row_items = []       
-        
-        for iid in selected:
-            for i, row in enumerate(tablerows):
-                if row.iid == iid:
-                    row_items.append([i, row])
-                    break
-
-        for index, item in reversed(row_items):
-            tablerows.remove(item)
-            tablerows.insert(index+1, item)
-
-        if self.filtered:
-            self.tablerows_filtered = tablerows
-        else:
-            self.tablerows = tablerows
-
-        # refresh the table data
-        self.unload_table_data()
-        self.load_table_data()
-
-    # COLUMN MOVEMENT
-
-    def move_column_left(self, event):
-        """Move column one position to the left
-        
-        Parameters:
-        
-            event (Event):
-                A window event.
-        """
-        eo = self._get_event_objects(event)
-        cid = str(eo.get('cid'))
-        displaycols = list(self.tableview.configure("displaycolumns"))
-        cols = list(self.tableview.configure("columns"))
-        if "#all" in displaycols:
-            displaycols = cols
-        old_index = displaycols.index(cid)
-        if old_index == 0:
-            return
-        new_index = old_index - 1
-        displaycols.insert(new_index, displaycols.pop(old_index))
-        self.tableview.configure(displaycolumns=displaycols)
-
-    def move_column_right(self, event):
-        """Move column one position to the right
-        
-        Parameters:
-        
-            event (Event):
-                A window event.
-        """
-        eo = self._get_event_objects(event)
-        cid = str(eo.get('cid'))
-        displaycols = list(self.tableview.configure("displaycolumns"))
-        cols = list(self.tableview.configure("columns"))
-        if "#all" in displaycols:
-            displaycols = cols
-        old_index = displaycols.index(cid)
-        if old_index == len(cols) - 1:
-            return
-        new_index = old_index + 1
-        displaycols.insert(new_index, displaycols.pop(old_index))
-        self.tableview.configure(displaycolumns=displaycols)
-
-    def move_column_to_first(self, event):
-        """Move column to leftmost position
-        
-        Parameters:
-        
-            event (Event):
-                A window event.
-        """
-        eo = self._get_event_objects(event)
-        cid = str(eo.get('cid'))
-        displaycols = list(self.tableview.configure("displaycolumns"))
-        cols = list(self.tableview.configure("columns"))
-        if "#all" in displaycols:
-            displaycols = cols
-        old_index = displaycols.index(cid)
-        if old_index == 0:
-            return
-        displaycols.insert(0, displaycols.pop(old_index))
-        self.tableview.configure(displaycolumns=displaycols)
-
-    def move_column_to_last(self, event):
-        """Move column to the rightmost position
-        
-        Parameters:
-        
-            event (Event):
-                A window event.
-        """
-        eo = self._get_event_objects(event)
-        cid = str(eo.get('cid'))
-        displaycols = list(self.tableview.configure("displaycolumns"))
-        cols = list(self.tableview.configure("columns"))
-        if "#all" in displaycols:
-            displaycols = cols
-        old_index = displaycols.index(cid)
-        if old_index == len(cols) - 1:
-            return
-        new_index = len(cols) - 1
-        displaycols.insert(new_index, displaycols.pop(old_index))
-        self.tableview.configure(displaycolumns=displaycols)    
-
-    # OTHER FORMATTING
-
-    def configure_table_stripes(self, stripecolor):
-        """Add stripes to even-numbered table rows as indicated by the
-        `stripecolor` of (background, foreground). Either element may be
-        specified as `None`, but both elements must be present.
-        
-        Parameters:
-
-            stripecolor (Tuple[str, str]):
-                A tuple of colors to apply to the table stripe. The
-                tuple represents (background, foreground).
-        """
-        if len(stripecolor) == 2:
-            self.stripecolor = stripecolor
-            bg, fg = stripecolor
-            kw = {}
-            if bg is not None:
-                kw["background"] = bg
-            if fg is not None:
-                kw["foreground"] = fg
-            self.tableview.tag_configure("striped", **kw)
-
-    def autofit_columns(self):
-        """Autofit all columns in the current view"""
-        f = font.nametofont("TkDefaultFont")
-        pad = utility.scale_size(self, 10)
-        col_widths = []
-
-        # measure header sizes
-        for col in self.tablecols:
-            width = f.measure(f'{col.headertext} {DOWNARROW}') + pad
-            col_widths.append(width)
-
-        for row in self.viewdata:
-            values = row.values
-            for i, value in enumerate(values):
-                old_width = col_widths[i]
-                new_width = f.measure(str(value)) + pad
-                width = max(old_width, new_width)
-                col_widths[i] = width
-        
-        for i, width in enumerate(col_widths):
-            self.tableview.column(i, width=width)
-
-    # COLUMN AND HEADER ALIGNMENT
-
-    def autoalign_columns(self):
-        """Align the columns and headers based on the data type of the
-        values. Text is left-aligned; numbers are right-aligned."""
-        values = self.tablerows[0].values
-        for i, value in enumerate(values):
-            if str(value).isnumeric():
-                self.tableview.column(i, anchor=E)
-                self.tableview.heading(i, anchor=E)
-
-    def align_column_left(self, event=None, cid=None):
-        """Left align the column text. This can be triggered by
-        either an event, or by passing in the `cid`, which is the index
-        of the column relative to the original data set.
-
-        Parameters:
-
-            event (Event):
-                An application event.
-
-            cid (int):
-                A unique column identifier; typically the index of the
-                column relative to the original dataset.
-        """
-        if event is not None:
-            eo = self._get_event_objects(event)
-            cid = eo.get("cid")
-        elif cid is None:
-            return
-        self.tableview.column(cid, anchor=W)
-
-    def align_column_right(self, event=None, cid=None):
-        """Right align the column text. This can be triggered by
-        either an event, or by passing in the `cid`, which is the index
-        of the column relative to the original data set.
-
-        Parameters:
-
-            event (Event):
-                An application event.
-
-            cid (int):
-                A unique column identifier; typically the index of the
-                column relative to the original dataset.
-        """
-        if event is not None:
-            eo = self._get_event_objects(event)
-            cid = eo.get("cid")
-        elif cid is None:
-            return
-        self.tableview.column(cid, anchor=E)
-
-    def align_column_center(self, event=None, cid=None):
-        """Center align the column text. This can be triggered by
-        either an event, or by passing in the `cid`, which is the index
-        of the column relative to the original data set.
-
-        Parameters:
-
-            event (Event):
-                An application event.
-
-            cid (int):
-                A unique column identifier; typically the index of the
-                column relative to the original dataset.
-        """
-        if event is not None:
-            eo = self._get_event_objects(event)
-            cid = eo.get("cid")
-        elif cid is None:
-            return
-        self.tableview.column(cid, anchor=CENTER)
-
-    def align_heading_left(self, event=None, cid=None):
-        """Left align the heading text. This can be triggered by
-        either an event, or by passing in the `cid`, which is the index
-        of the heading relative to the original data set.
-
-        Parameters:
-
-            event (Event):
-                An application event.
-
-            cid (int):
-                A unique heading identifier; typically the index of the
-                heading relative to the original dataset.
-        """
-        if event is not None:
-            eo = self._get_event_objects(event)
-            cid = eo.get("cid")
-        elif cid is None:
-            return
-        self.tableview.heading(cid, anchor=W)
-
-    def align_heading_right(self, event=None, cid=None):
-        """Right align the heading text. This can be triggered by
-        either an event, or by passing in the `cid`, which is the index
-        of the heading relative to the original data set.
-
-        Parameters:
-
-            event (Event):
-                An application event.
-
-            cid (int):
-                A unique heading identifier; typically the index of the
-                heading relative to the original dataset.
-        """
-        if event is not None:
-            eo = self._get_event_objects(event)
-            cid = eo.get("cid")
-        elif cid is None:
-            return
-        self.tableview.heading(cid, anchor=E)
-
-    def align_heading_center(self, event=None, cid=None):
-        """Center align the heading text. This can be triggered by
-        either an event, or by passing in the `cid`, which is the index
-        of the heading relative to the original data set.
-
-        Parameters:
-
-            event (Event):
-                An application event.
-
-            cid (int):
-                A unique heading identifier; typically the index of the
-                heading relative to the original dataset.
-        """
-        if event is not None:
-            eo = self._get_event_objects(event)
-            cid = eo.get("cid")
-        elif cid is None:
-            return
-        self.tableview.heading(cid, anchor=CENTER)       
-
-    # WIDGET BINDING
-
-    def _set_widget_binding(self):
-        """Setup the widget binding"""
-        self.tableview.bind("<Double-Button-1>", self._header_double_leftclick)
-        self.tableview.bind("<Button-1>", self._header_leftclick)
-        if self.tk.call('tk', 'windowingsystem') == 'aqua':
-            sequence = "<Button-2>"
-        else:
-            sequence = "<Button-3>"
-        self.tableview.bind(sequence, self._table_rightclick)
-
-        # add trace to track pagesize changes
-        # self.pagesize.trace_add('write', self._trace_pagesize)
-
-    def _select_pagesize(self, event):
-        cbo: ttk.Combobox = self.nametowidget(event.widget)
-        cbo.select_clear()
-        self.goto_first_page()
-
-    def _trace_pagesize(self, *_):
-        """Callback for changes to page size"""
-        #pagesize = self.pagesize.get()
-        #self.tableview.configure(height=pagesize)
-        self.goto_first_page()
-
-    def _header_double_leftclick(self, event):
-        """Callback for double-click events on the tableview header"""
-        region = self.tableview.identify_region(event.x, event.y)
-        if region == "separator":
-            self.autofit_columns()
-
-    def _header_leftclick(self, event):
-        """Callback for left-click events"""
-        region = self.tableview.identify_region(event.x, event.y)
-        if region == "heading":
-            # col = self.tableview.identify_column(event.x)
-            # cid = int(self.tableview.column(col, "id"))
-            self.sort_column_data(event)
-
-    def _table_rightclick(self, event):
-        """Callback for right-click events"""
-        region = self.tableview.identify_region(event.x, event.y)
-        if region == "heading":
-            self.rightclickmenu_head.post(event)
-        elif region != "separator":
-            self.rightclickmenu_cell.post(event)
-
-
-class TableColumn:
-    """Represents a column in a Tableview object - INTERNAL"""
-
-    def __init__(
-        self,
-        table,
-        cid,
-        text,
-        image="",
-        command="",
-        anchor=W,
-        width=200,
-        minwidth=20,
-        maxwidth=400,
-        stretch=False,
-    ):
-        """
-        Parameters:
-
-            table (Treeview):
-                The internal Treeview object of the Tableview.
-
-            cid (str):
-                The column id.
-
-            text (str):
-                The header text.
-
-            image (PhotoImage):
-                An image that is displayed to the left of the header text.
-
-            command (Callable):
-                A function called whenever the header button is clicked.
-
-            anchor (str):
-                The position of the header text within the header. One
-                of "e", "w", "center".
-
-            width (int):
-                Specifies the width of the column in pixels.
-
-            minwidth (int):
-                Specifies the minimum width of the column in pixels.
-
-            maxwidth (int):
-                Specifies the maximum width of the column in pixels.
-
-            stretch (bool):
-                Specifies whether or not the column width should be
-                adjusted whenever the widget is resized or the user
-                drags the column separator.
-        """
-        self.cid = cid
-        self.headertext = text
-        self.table: ttk.Treeview = table
-        self.colmaxwidth = maxwidth
-        self.sort = ASCENDING
-        self.tableview = None
-        self.hbar = None
-        self.table.column(
-            self.cid,
-            width=width,
-            minwidth=minwidth,
-            stretch=stretch,
-            anchor=anchor,
-        )
-        self.table.heading(
-            self.cid,
-            text=text,
-            anchor=anchor,
-            image=image,
-            command=command,
-        )
-
-
-class TableRow:
-    """Represents a row in a Tableview object - INTERNAL"""
-
-    def __init__(self, table, values):
-        """
-        Parameters:
-
-            table (Treeview):
-                The inner Treeview object within the Tableview
-
-            values (List[Any, ...]):
-                A list of values to display in the row
-        """
-        self.table: ttk.Treeview = table
-        self.values = values
-        self.iid = None
-
-    def show(self, striped=False):
-        """Show the row in the data table view"""
-        if self.iid is None:
-            self.build_row()
-        self.table.reattach(self.iid, "", END)
-
-        # remove existing stripes
-        tags = list(self.table.item(self.iid, "tags"))
-        # TODO where is the `table.tag_remove` method?
-        try:
-            tags.remove("striped")
-        except ValueError:
-            pass
-
-        # add stripes (if needed)
-        if striped:
-            tags.append("striped")
-        self.table.item(self.iid, tags=tags)
-
-    def hide(self):
-        """Remove the row from the data table view"""
-        self.table.detach(self.iid)
-
-    # def cell_configure(self, index, value):
-    #     """Modify the value of a specific cell"""
-    #     self.values[index] = value
-    #     self.table.item(self.iid, values=self.values)
-
-    def build_row(self):
-        """Create the row object in the `Treeview` and capture
-        the resulting item id (iid).
-        """
-        self.iid = self.table.insert("", END, values=self.values)
-
-
-class TableCellRightClickMenu(tk.Menu):
-    """A right-click menu object for the tableview cells - INTERNAL"""
-
-    def __init__(self, master: Tableview):
-        """
-        Parameters:
-
-            master (Tableview):
-                The parent object
-        """
-        super().__init__(master, tearoff=False)
-        self.tableview: Tableview = master
-        self.treeview: ttk.Treeview = master.tableview
-        self.cid = None
-        self.iid = None
-
-        config = {
-            "sortascending": {
-                "label": "⬆  Sort Ascending",
-                "command": self.sort_column_ascending,
-            },
-            "sortdescending": {
-                "label": "⬇  Sort Descending",
-                "command": self.sort_column_descending,
-            },
-            "clearfilter": {
-                "label": "Clear filter",
-                "command": self.tableview.clear_table_filters,
-            },
-            "filterbyvalue": {
-                "label": "Filter by cell's value",
-                "command": self.filter_to_cell_value,
-            },
-            "hiderows": {
-                "label": "Hide select rows",
-                "command": self.hide_selected_rows,
-            },
-            "showrows": {
-                "label": "Show only select rows",
-                "command": self.filter_to_selected_rows,
-            },
-            "exportall": {
-                "label": "Export all records",
-                "command": self.export_all_records,
-            },
-            "exportpage": {
-                "label": "Export current page",
-                "command": self.export_current_page,
-            },
-            "exportselection": {
-                "label": "Export current selection",
-                "command": self.export_current_selection,
-            },
-            "exportfiltered": {
-                "label": "Export records in filter",
-                "command": self.export_records_in_filter,
-            },
-            "moveup": {
-                "label": "↑ Move up", 
-                "command": self.move_row_up
-            },
-            "movedown": {
-                "label": "↓ Move down",
-                "command": self.move_row_down,
-            },
-            "movetotop": {
-                "label": "⤒ Move to top",
-                "command": self.move_row_to_top,
-            },
-            "movetobottom": {
-                "label": "⤓ Move to bottom",
-                "command": self.move_row_to_bottom,
-            },
-            "alignleft": {
-                "label": "◧  Align left",
-                "command": self.align_column_left,
-            },
-            "aligncenter": {
-                "label": "◫  Align center",
-                "command": self.align_column_center,
-            },
-            "alignright": {
-                "label": "◨  Align right",
-                "command": self.align_column_right,
-            },
-        }
-        sort_menu = tk.Menu(self, tearoff=False)
-        sort_menu.add_command(cnf=config["sortascending"])
-        sort_menu.add_command(cnf=config["sortdescending"])
-        self.add_cascade(menu=sort_menu, label="⇅ Sort")
-
-        filter_menu = tk.Menu(self, tearoff=False)
-        filter_menu.add_command(cnf=config["clearfilter"])
-        filter_menu.add_separator()
-        filter_menu.add_command(cnf=config["filterbyvalue"])
-        filter_menu.add_command(cnf=config["hiderows"])
-        filter_menu.add_command(cnf=config["showrows"])
-        self.add_cascade(menu=filter_menu, label="⧨  Filter")
-
-        export_menu = tk.Menu(self, tearoff=False)
-        export_menu.add_command(cnf=config["exportall"])
-        export_menu.add_command(cnf=config["exportpage"])
-        export_menu.add_command(cnf=config["exportselection"])
-        export_menu.add_command(cnf=config["exportfiltered"])
-        self.add_cascade(menu=export_menu, label="↔  Export")
-
-        move_menu = tk.Menu(self, tearoff=False)
-        move_menu.add_command(cnf=config["moveup"])
-        move_menu.add_command(cnf=config["movedown"])
-        move_menu.add_command(cnf=config["movetotop"])
-        move_menu.add_command(cnf=config["movetobottom"])
-        self.add_cascade(menu=move_menu, label="⇵  Move")
-
-        align_menu = tk.Menu(self, tearoff=False)
-        align_menu.add_command(cnf=config["alignleft"])
-        align_menu.add_command(cnf=config["aligncenter"])
-        align_menu.add_command(cnf=config["alignright"])
-        self.add_cascade(menu=align_menu, label="↦  Align")
-
-    def post(self, event):
-        """Display the menu below the selected cell.
-
-        Parameters:
-
-            event (Event):
-                The click event that triggers menu.
-        """
-        # capture the column and item that invoked the menu
-        self.event = event
-        iid = self.treeview.identify_row(event.y)
-        col = self.treeview.identify_column(event.x)
-
-        # show the menu below the invoking cell
-        rootx = self.treeview.winfo_rootx()
-        rooty = self.treeview.winfo_rooty()
-        bbox = self.treeview.bbox(iid, col)
-        super().post(rootx + bbox[0], rooty + bbox[1] + bbox[3])
-
-    def sort_column_ascending(self):
-        """Sort the column in ascending order."""
-        self.tableview.sort_column_data(self.event, sort=ASCENDING)
-
-    def sort_column_descending(self):
-        """Sort the column in descending order."""
-        self.tableview.sort_column_data(self.event, sort=DESCENDING)
-
-    def filter_to_cell_value(self):
-        """Hide all records except for records where the current
-        column exactly matches the current cell value."""
-        self.tableview.filter_column_to_value(self.event)
-
-    def filter_to_selected_rows(self):
-        """Hide all records except for the selected rows."""
-        self.tableview.filter_to_selected_rows()
-
-    def export_all_records(self):
-        """Export all records to a csv file"""
-        self.tableview.export_all_records()
-
-    def export_current_page(self):
-        """Export records on current page"""
-        self.tableview.export_current_page()
-
-    def export_current_selection(self):
-        """Export rows currently selected"""
-        self.tableview.export_current_selection()
-
-    def export_records_in_filter(self):
-        """Export rows currently filtered"""
-        self.tableview.export_records_in_filter()
-
-    def hide_selected_rows(self):
-        """Hide the selected rows"""
-        self.tableview.hide_selected_rows()
-
-    def move_row_to_top(self):
-        """Move the row to the top of the data set"""
-        self.tableview.move_selected_rows_to_top()
-
-    def move_row_to_bottom(self):
-        """Move the row to the bottom of the dataset"""
-        self.tableview.move_selected_rows_to_bottom()
-
-    def move_row_up(self):
-        """Move the selected above the previous sibling"""
-        self.tableview.move_selected_row_up()
-
-    def move_row_down(self):
-        """Move the selected row below the next sibling"""
-        self.tableview.move_row_down()
-
-    def align_column_left(self):
-        "Left align the column text"
-        self.tableview.align_column_left(self.event)
-
-    def align_column_right(self):
-        """Right align the column text"""
-        self.tableview.align_column_right(self.event)
-
-    def align_column_center(self):
-        """Center align the column text"""
-        self.tableview.align_column_center(self.event)
-
-
-class TableHeaderRightClickMenu(tk.Menu):
-    """A right-click menu object for the tableview header - INTERNAL"""
-
-    def __init__(self, master: Tableview):
-        """
-        Parameters:
-
-            master (Tableview):
-                The parent object
-        """
-        super().__init__(master, tearoff=False)
-        self.tableview: Tableview = self.master
-        self.table: ttk.Treeview = master.tableview
-        self.event = None
-
-        # HIDE & SHOW
-        show_menu = tk.Menu(self, tearoff=False)
-        for column in self.master.tablecols:
-            variable = f"column_{column.cid}"
-            self.table.setvar(variable, True)
-            show_menu.add_checkbutton(
-                label=column.headertext,
-                command=lambda w=column: self.toggle_columns(w.cid),
-                variable=variable,
-                indicatoron=True,
-                onvalue=True,
-                offvalue=False,
-            )
-        self.add_cascade(menu=show_menu, label="±  Columns")
-
-        config = {
-            'movetoright': {
-                "label": "→  Move to right", 
-                "command": self.move_column_right
-            },
-            'movetoleft': {
-                "label": "←  Move to left", 
-                "command": self.move_column_left
-            },
-            'movetofirst': {
-                "label": "⇤  Move to first", 
-                "command": self.move_column_to_first
-            },
-            'movetolast': {
-                "label": "⇥  Move to last", 
-                "command": self.move_column_to_last
-            },
-            'alignleft': {
-                "label": "◧  Align left", 
-                "command": self.align_heading_left
-            },
-            'alignright': {
-                "label": "◨  Align right", 
-                "command": self.align_heading_right                
-            },
-            'aligncenter': {
-                "label": "◫  Align center", 
-                "command": self.align_heading_center
-            }
-        }
-
-        # MOVE MENU
-        move_menu = tk.Menu(self, tearoff=False)
-        move_menu.add_command(cnf=config['movetoleft'])
-        move_menu.add_command(cnf=config['movetoright'])
-        move_menu.add_command(cnf=config['movetofirst'])
-        move_menu.add_command(cnf=config['movetolast'])
-        self.add_cascade(menu=move_menu, label="⇄  Move")
-
-        align_menu = tk.Menu(self, tearoff=False)
-        align_menu.add_command(cnf=config["alignleft"])
-        align_menu.add_command(cnf=config["aligncenter"])
-        align_menu.add_command(cnf=config["alignright"])
-        self.add_cascade(menu=align_menu, label="↦  Align")
-
-    def post(self, event):
-        # capture the column and item that invoked the menu
-        self.event = event
-
-        # show the menu below the invoking cell
-        rootx = self.table.winfo_rootx()
-        rooty = self.table.winfo_rooty()
-        super().post(rootx + event.x, rooty + event.y + 10)
-
-    def toggle_columns(self, cid):
-        """Toggles the visibility of the selected column"""
-        variable = f"column_{cid}"
-        toggled = self.getvar(variable)
-        if toggled:
-            self.tableview.show_selected_column(cid=cid)
-        else:
-            self.tableview.hide_selected_column(cid=cid)
-
-    def move_column_left(self):
-        """Move column one position to the left"""
-        self.tableview.move_column_left(self.event)
-
-    def move_column_right(self):
-        """Move column on position to the right"""
-        self.tableview.move_column_right(self.event)
-
-    def move_column_to_first(self):
-        """Move column to leftmost position"""
-        self.tableview.move_column_to_first(self.event)
-
-    def move_column_to_last(self):
-        """Move column to rightmost position"""
-        self.tableview.move_column_to_last(self.event)
-
-    def align_heading_left(self):
-        """Left align the column header"""
-        self.tableview.align_heading_left(self.event)
-
-    def align_heading_right(self):
-        """Right align the column header"""
-        self.tableview.align_heading_right(self.event)
-
-    def align_heading_center(self):
-        """Center align the column header"""
-        self.tableview.align_heading_center(self.event)
